wip: Implemented clippy suggestions

+++ src/utils.rs
@@ -78,7 +78,7 @@ pub fn is_valid_coord(coord: u8) -> bool {
 }
 
 pub fn compact_pos(rank: u8, file: u8) -> u8 {
-    (rank as u8) << 4 | (file as u8)
+    rank << 4 | file
 }
 
 pub fn unpack_pos<T: From<u8>, V: Into<u8>>(pos: V) -> (T, T) {
